test(loader): give the backend-DLL contract its own owner
test_cuda_env.py had grown to cover two unrelated contracts: the COLI_CUDA
environment semantics it was written for and the backend-DLL selection added
for Windows HIP. Split the second into its own module so each file states one
thing and the loader tests no longer pay for the other module's import-time
probe of the resident colibri binary.

- move LoaderBackendSelectionTest to test_backend_loader.py
- preserve every test name, executable code line and assertion
- retain isolated CUDA_DLL and HIP_DLL host builds and sandboxes
- keep test_cuda_env.py focused on COLI_CUDA startup semantics
- use private fixture helpers instead of importing a module with import-time
  binary and GPU probes
- remove the now-unused shutil import from test_cuda_env.py

No production behaviour or test semantics change. The suite remains at 277
Python tests with 40 skipped, split as seven environment tests and four loader
tests for the extracted owner.
